@@ -35,18 +35,18 @@
   },
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular-devkit/build-angular": "~0.803.29",
-    "@angular-devkit/build-ng-packagr": "~0.803.29",
-    "@angular/animations": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.901.12",
+    "@angular-devkit/build-ng-packagr": "~0.901.12",
+    "@angular/animations": "^9.1.12",
     "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
+    "@angular/common": "^9.1.12",
+    "@angular/compiler": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/forms": "^9.1.12",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/platform-browser": "^9.1.12",
+    "@angular/platform-browser-dynamic": "^9.1.12",
+    "@angular/router": "^9.1.12",
     "@google/markerclusterer": "^2.0.0",
     "@nebular/auth": "^4.6.0",
     "@nebular/bootstrap": "^4.6.0",
@@ -81,7 +81,7 @@
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
     "nebular-icons": "^1.1.0",
-    "ng-packagr": "^5.4.0",
+    "ng-packagr": "^9.0.0",
     "ng-pick-datetime": "^7.0.0",
     "ng2-completer": "^2.0.8",
     "ng2-pdf-viewer": "^5.3.4",
@@ -97,24 +97,24 @@
     "stream": "0.0.2",
     "timers": "^0.1.1",
     "topojson-client": "^3.1.0",
-    "tslib": "^1.13.0",
+    "tslib": "^1.10.0",
     "typeface-exo": "0.0.22",
     "web-animations-js": "^2.3.2",
     "xlsx": "^0.14.3",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
     "@angular-devkit/core": "^10.0.4",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular/cli": "^9.1.12",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/language-service": "9.1.12",
     "@compodoc/compodoc": "^1.1.10",
     "@fortawesome/fontawesome-free": "^5.14.0",
     "@types/d3": "^5.7.2",
     "@types/jasmine": "2.5.54",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.90",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
     "jasmine-core": "2.6.4",
@@ -132,6 +132,6 @@
     "stylelint": "^9.10.1",
     "ts-node": "3.2.2",
     "tslint": "5.7.0",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3"
   }
 }
