@@ -15,7 +15,7 @@ declare var google: any;
 })
 export class VehicleStatesComponent implements OnInit {
   map: any;
-  @ViewChild('map', { static: false }) mapElement: ElementRef;
+  @ViewChild('map') mapElement: ElementRef;
   marker: any;
   eventInfo = null;
   infoWindow = null;
